Refactor. Move infra type code to Infrastructure assembly and clean-up App project. Part 1 of 2.
