+++ src/EM.HostsManager.App/GlobalUsings.cs
@@ -4,10 +4,5 @@ global using System;
 global using System.Diagnostics;
 global using System.Diagnostics.CodeAnalysis;
 global using System.IO;
-global using System.Linq;
 global using System.Reflection;
-global using System.Runtime.InteropServices;
-global using System.Security.Principal;
-global using System.Threading;
 global using System.Windows.Forms;
-global using System.Collections.Generic;
