+++ src/EM.HostsManager.App/Program.BuildDate.cs
@@ -1,8 +1,8 @@
 ﻿//
-// Copyright © 2021-2022 Enda Mullally.
+// Copyright © 2021-2023 Enda Mullally.
 //
 
-using EM.HostsManager.App.Attributes;
+using EM.HostsManager.Infrastructure.Attributes;
 
 // BuildDate
 [assembly: BuildDate("BUILD-DATE-ATTRIBUTE")]
