+++ src/EM.HostsManager.Infrastructure/GlobalUsings.cs
@@ -1,0 +1,7 @@
+// Global using directives
+
+global using System.Reflection;
+global using System.Diagnostics;
+global using System.Security.Principal;
+global using System.Runtime.InteropServices;
+global using Microsoft.Win32;
