+++ src/EM.HostsManager.Infrastructure/Version/AppVersion.cs
@@ -1,10 +1,10 @@
 ﻿//
-// Copyright © 2021-2022 Enda Mullally.
+// Copyright © 2021-2023 Enda Mullally.
 //
 
-using EM.HostsManager.App.Attributes;
+using EM.HostsManager.Infrastructure.Attributes;
 
-namespace EM.HostsManager.App.Version;
+namespace EM.HostsManager.Infrastructure.Version;
 
 public class AppVersion
 {
