+++ src/EM.HostsManager.sln
@@ -5,6 +5,8 @@ VisualStudioVersion = 17.3.32922.545
 MinimumVisualStudioVersion = 10.0.40219.1
 Project("{9A19103F-16F7-4668-BE54-9A1E7A4F7556}") = "EM.HostsManager.App", "EM.HostsManager.App\EM.HostsManager.App.csproj", "{014D346F-73EC-4060-9095-D8028E0FBD67}"
 EndProject
+Project("{FAE04EC0-301F-11D3-BF4B-00C04F79EFBC}") = "EM.HostsManager.Infrastructure", "EM.HostsManager.Infrastructure\EM.HostsManager.Infrastructure.csproj", "{EFC8F105-4534-4E19-BBA1-5DF2E5004172}"
+EndProject
 Global
 	GlobalSection(SolutionConfigurationPlatforms) = preSolution
 		Debug|Any CPU = Debug|Any CPU
@@ -15,6 +17,10 @@ Global
 		{014D346F-73EC-4060-9095-D8028E0FBD67}.Debug|Any CPU.Build.0 = Debug|Any CPU
 		{014D346F-73EC-4060-9095-D8028E0FBD67}.Release|Any CPU.ActiveCfg = Release|Any CPU
 		{014D346F-73EC-4060-9095-D8028E0FBD67}.Release|Any CPU.Build.0 = Release|Any CPU
+		{EFC8F105-4534-4E19-BBA1-5DF2E5004172}.Debug|Any CPU.ActiveCfg = Debug|Any CPU
+		{EFC8F105-4534-4E19-BBA1-5DF2E5004172}.Debug|Any CPU.Build.0 = Debug|Any CPU
+		{EFC8F105-4534-4E19-BBA1-5DF2E5004172}.Release|Any CPU.ActiveCfg = Release|Any CPU
+		{EFC8F105-4534-4E19-BBA1-5DF2E5004172}.Release|Any CPU.Build.0 = Release|Any CPU
 	EndGlobalSection
 	GlobalSection(SolutionProperties) = preSolution
 		HideSolutionNode = FALSE
